It's #1548 but without conflicts (it was easier for me to cherry pick some of the commits and write rest from scratch)

|        Method |      Args |    Before |     After |    Before |     After |
|-------------- |---------- |----------:|----------:|----------:|----------:|
|  DefaultsSync | String[0] | 27.885 us | 26.447 us |  11,160 B |  10,872 B |
| DefaultsAsync | String[0] | 28.015 us | 27.131 us |  11,160 B |  10,872 B |
|   MinimalSync | String[0] |  2.334 us |  2.240 us |   4,480 B |   4,224 B |
|  MinimalAsync | String[0] |  2.428 us |  2.345 us |   4,480 B |   4,224 B |
|  DefaultsSync | String[4] | 31.778 us | 31.072 us |  14,792 B |  14,304 B |
| DefaultsAsync | String[4] | 32.816 us | 31.284 us |  14,792 B |  14,304 B |
|   MinimalSync | String[4] |  4.898 us |  4.530 us |   8,112 B |   7,656 B |
|  MinimalAsync | String[4] |  5.080 us |  4.659 us |   8,112 B |   7,656 B |